Revamp date system; change type system
Change storage system so that not everything is stored as a LinkedText
object; only autolinked items are stored as LinkedText. Plain text items
are stored as strings (str), and timestamps are stored as datetime objects.

Add date parsing functionality so that the user's text timestamp input
can be converted to a datetime object. Use the 'dateutil' library for
parsing (dateutil.parser.parse()).